Add files via upload
This is updated to represent what we have talked about in class and I can't figure out how to setup foreign keys they need to be setup in the communication table to connect to the other three tables. 

Look in Docs at the phase 2 ERD as a basic blueprint remember that Mrs. Angela told us to remove the connection between the student and activities table which is why it is not there.
I also removed state, city, and zip from students. I added a "local" colomn that is tinyint (boolean) that we can use to track if students are in or out of the country. I also added a "ststus" which is if they should continue to reach out to this student or not. And finally an "activitiesCounter" which counts down from 30 and will reset once a communication has been made with the student or if they are deactivated it will stay on 30 permanently. (This stuff we need to code. Im just explaining how it should work.

In the activity table I added a name section and made the description and the date nullable so that they could add basic activities like "sent email", "met in office" things they may not want to date.

In the users table I gave them "email" (so we can send emails to them) and a "privledged rights" colomn so we can allow only certain teachers to have more control over the applicationthan others for deletiong stuff later for example.